Removed some potential problems with SimpleQueue. I didn't prove that it was buggy, but this was certainly very fragile code. (Modyfing a queue through an iterator intended for a list is a 'nice' way to create incorrect afterlast pointers.) + cleanup of some warnings in debug-mode DSS. + allows dp-bench to compile (but not yet run.) For how long has this stuff been broken?

+++ platform/dss/src/msl_transObj.cc
@@ -73,7 +73,7 @@ namespace _msl_internal{
   const int TRAILER = 1;                       // control flag
 
   void BufferedTransObj::marshal(MsgCnt *msgC, int acknum) {
-    int framesize;
+    size_t framesize;
     BYTE* pos;
 
     dssLog(DLL_TOO_MUCH,"TRANSOBJ      (%p): Marshal  ack:%d cont:%d",
